Python/requirements.txt
pyzmq
opencv-python
